Fix null object bug
When started, the program would complain about a null
pointer exception. This exception occurred because the example OpModes
would supply a null hardwareMap object into the constructor of
FourWheelRobot. The hardwareMap object was null because I referenced it
in the constructor of the OpMode, which is before the hardwareMap object
is initialized.

To fix this, add a null pointer check into the constructor of
FourWheelRobot,
and move the FourWheelRobot initializations in the example OpMode
classes into the runOpMode method.